requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2
django-allauth==0.40.0
django-cors-headers==3.2.1
django-countries==5.5
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-rest-auth==0.9.5
django-tinymce4-lite==1.7.5
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.2
mccabe==0.6.1
mysqlclient==1.4.6
oauthlib==3.1.0
pep8==1.7.1
Pillow==6.1.0
pipenv==2018.11.26
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.2.0
rope==0.14.0
six==1.12.0
sqlparse==0.2.4
stripe==2.41.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
wrapt==1.11.2